Factor out lambda-io.tsx.
